src/requirements.txt
huggingface_hub
matplotlib
numpy
pandas
python-dotenv
PyYAML
rank_bm25
scikit_learn
sentence_transformers
torch
tqdm
transformers
vllm

